requirements.prod.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.0.2
flask-migrate==4.0.4
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
werkzeug==3.0.4
google-generativeai
requests>=2.28.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
